+++ packages/react/src/themes/teams-high-contrast/components/Chat/chatMessageVariables.ts
@@ -7,7 +7,6 @@ export default (siteVars: any): Partial<ChatMessageVariables> => {
     authorColor: siteVars.colors.white,
     contentColor: siteVars.colors.white,
     color: siteVars.colors.white,
-    linkColor: siteVars.accessibleYellow,
     border: `1px solid ${siteVars.colors.white}`,
     hasMentionColor: siteVars.accessibleYellow,
     hasMentionNubbinColor: siteVars.accessibleYellow,